[#936] Turn DisplayCAL.icc_profile into a package (item 1: constants, codecs, structures)
`DisplayCAL/icc_profile.py` is a 10.8k line module with no internal
boundaries. This starts the split into `DisplayCAL/icc_profile/`, moving
the module docstring/exports-holding file to `icc_profile/__init__.py`
and pulling out the three lowest-level, zero-internal-dependency groups
into their own modules:

- `constants.py`: `TAGS`, `COLORANTS`, `ILLUMINANTS`, `MANUFACTURERS`,
  `PROFILE_CLASS`, `TECH`, `CMMS`, `ENCODINGS`, `GAMUT_VOLUME_*`, etc.
- `codecs.py`: `s15Fixed16Number`/`uInt*Number`/`dateTimeNumber`/
  `videoCardGamma` (+`_tohex` pairs), `hexrepr`, `legacy_PCSLab_*`.
- `structures.py`: `ADict`, `AODict`, `DictList`, `DictListItem`,
  `CRInterpolation`.

`icc_profile/__init__.py` imports every name back explicitly, so none of
the ~37 modules or ~15 test files importing from `DisplayCAL.icc_profile`
need to change. `videoCardGamma()` uses a deferred import of
`VideoCardGammaFormulaType`/`VideoCardGammaTableType` to avoid a circular
import with the tag classes, which still live in `__init__.py` pending a
later punch-list item.

`LazyLoadTagAODict` stayed in `__init__.py` rather than moving to
`structures.py` as originally sketched in #936, it's tightly coupled to
`ICCProfileTag` and the tag-signature registries, not a zero-dependency
structure, so it belongs with the tag classes in a later item instead.
