+++ DisplayCAL/icc_profile/__init__.py
@@ -120,375 +120,61 @@ if TYPE_CHECKING:
         from typing_extensions import Self
 
 
-# Gamut volumes in cubic colorspace units (L*a*b*) as reported by Argyll's
-# iccgamut
-GAMUT_VOLUME_SRGB = 833675.435316  # rel. col.
-GAMUT_VOLUME_ADOBERGB = 1209986.014983  # rel. col.%
-GAMUT_VOLUME_SMPTE431_P3 = 1176953.485921  # rel. col.
-
-# http://msdn.microsoft.com/en-us/library/dd371953%28v=vs.85%29.aspx
-COLOR_PROFILE_SUBTYPE = {
-    "NONE": 0x0000,
-    "RGB_WORKING_SPACE": 0x0001,
-    "PERCEPTUAL": 0x0002,
-    "ABSOLUTE_COLORIMETRIC": 0x0004,
-    "RELATIVE_COLORIMETRIC": 0x0008,
-    "SATURATION": 0x0010,
-    "CUSTOM_WORKING_SPACE": 0x0020,
-}
-
-# http://msdn.microsoft.com/en-us/library/dd371955%28v=vs.85%29.aspx (wrong)
-# http://msdn.microsoft.com/en-us/library/windows/hardware/ff546018%28v=vs.85%29.aspx (ok)  # noqa: E501
-COLOR_PROFILE_TYPE = {"ICC": 0, "DMP": 1, "CAMP": 2, "GMMP": 3}
-
-WCS_PROFILE_MANAGEMENT_SCOPE = {"SYSTEM_WIDE": 0, "CURRENT_USER": 1}
-
-ERROR_PROFILE_NOT_ASSOCIATED_WITH_DEVICE = 2015
-ERROR_SUCCESS = 0
-
-DEBUG = False
-
-ENC, FS_ENC = get_encodings()
-
-CMMS = {
-    b"argl": "ArgyllCMS",
-    b"ADBE": "Adobe",
-    b"ACMS": "Agfa",
-    b"Agfa": "Agfa",
-    b"APPL": "Apple",
-    b"appl": "Apple",
-    b"CCMS": "ColorGear",
-    b"UCCM": "ColorGear Lite",
-    b"DL&C": "Digital Light & Color",
-    b"EFI ": "EFI",
-    b"FF  ": "Fuji Film",
-    b"HCMM": "Harlequin RIP",
-    b"LgoS": "LogoSync",
-    b"HDM ": "Heidelberg",
-    b"Lino": "Linotype",
-    b"lino": "Linotype",
-    b"lcms": "Little CMS",
-    b"KCMS": "Kodak",
-    b"MCML": "Konica Minolta",
-    b"MSFT": "Microsoft",
-    b"SIGN": "Mutoh",
-    b"RGMS": "DeviceLink",
-    b"SICC": "SampleICC",
-    b"32BT": "the imaging factory",
-    b"WTG ": "Ware to Go",
-    b"zc00": "Zoran",
-}
-
-ENCODINGS = {
-    "mac": {
-        141: "africaans",
-        36: "albanian",
-        85: "amharic",
-        12: "arabic",
-        51: "armenian",
-        68: "assamese",
-        134: "aymara",
-        49: "azerbaijani-cyrllic",
-        50: "azerbaijani-arabic",
-        129: "basque",
-        67: "bengali",
-        137: "dzongkha",
-        142: "breton",
-        44: "bulgarian",
-        77: "burmese",
-        46: "byelorussian",
-        78: "khmer",
-        130: "catalan",
-        92: "chewa",
-        33: "simpchinese",
-        19: "tradchinese",
-        18: "croatian",
-        38: "czech",
-        7: "danish",
-        4: "dutch",
-        0: "roman",
-        94: "esperanto",
-        27: "estonian",
-        30: "faeroese",
-        31: "farsi",
-        13: "finnish",
-        34: "flemish",
-        1: "french",
-        140: "galician",
-        144: "scottishgaelic",
-        145: "manxgaelic",
-        52: "georgian",
-        2: "german",
-        14: "greek-monotonic",
-        148: "greek-polytonic",
-        133: "guarani",
-        69: "gujarati",
-        10: "hebrew",
-        21: "hindi",
-        26: "hungarian",
-        15: "icelandic",
-        81: "indonesian",
-        143: "inuktitut",
-        35: "irishgaelic",
-        146: "irishgaelic-dotsabove",
-        3: "italian",
-        11: "japanese",
-        138: "javaneserom",
-        73: "kannada",
-        61: "kashmiri",
-        48: "kazakh",
-        90: "kiryarwanda",
-        54: "kirghiz",
-        91: "rundi",
-        23: "korean",
-        60: "kurdish",
-        79: "lao",
-        131: "latin",
-        28: "latvian",
-        24: "lithuanian",
-        43: "macedonian",
-        93: "malagasy",
-        83: "malayroman-latin",
-        84: "malayroman-arabic",
-        72: "malayalam",
-        16: "maltese",
-        66: "marathi",
-        53: "moldavian",
-        57: "mongolian",
-        58: "mongolian-cyrillic",
-        64: "nepali",
-        9: "norwegian",
-        71: "oriya",
-        87: "oromo",
-        59: "pashto",
-        25: "polish",
-        8: "portuguese",
-        70: "punjabi",
-        132: "quechua",
-        37: "romanian",
-        32: "russian",
-        29: "sami",
-        65: "sanskrit",
-        42: "serbian",
-        62: "sindhi",
-        76: "sinhalese",
-        39: "slovak",
-        40: "slovenian",
-        88: "somali",
-        6: "spanish",
-        139: "sundaneserom",
-        89: "swahili",
-        5: "swedish",
-        82: "tagalog",
-        55: "tajiki",
-        74: "tamil",
-        135: "tatar",
-        75: "telugu",
-        22: "thai",
-        63: "tibetan",
-        86: "tigrinya",
-        147: "tongan",
-        17: "turkish",
-        56: "turkmen",
-        136: "uighur",
-        45: "ukrainian",
-        20: "urdu",
-        47: "uzbek",
-        80: "vietnamese",
-        128: "welsh",
-        41: "yiddish",
-    }
-}
-
-COLORANTS = {
-    0: {"description": "unknown", "channels": ()},
-    1: {
-        "description": "ITU-R BT.709",
-        "channels": ((0.64, 0.33), (0.3, 0.6), (0.15, 0.06)),
-    },
-    2: {
-        "description": "SMPTE RP145-1994",
-        "channels": ((0.63, 0.34), (0.31, 0.595), (0.155, 0.07)),
-    },
-    3: {
-        "description": "EBU Tech.3213-E",
-        "channels": ((0.64, 0.33), (0.29, 0.6), (0.15, 0.06)),
-    },
-    4: {
-        "description": "P22",
-        "channels": ((0.625, 0.34), (0.28, 0.605), (0.155, 0.07)),
-    },
-}
-
-GEOMETRY = {0: "unknown", 1: "0/45 or 45/0", 2: "0/d or d/0"}
-
-ILLUMINANTS = {
-    0: "unknown",
-    1: "D50",
-    2: "D65",
-    3: "D93",
-    4: "F2",
-    5: "D55",
-    6: "A",
-    7: "E",
-    8: "F8",
-}
-
-OBSERVERS = {0: "unknown", 1: "CIE 1931", 2: "CIE 1964"}
-
-MANUFACTURERS = {
-    b"ADBE": "Adobe Systems Incorporated",
-    b"APPL": "Apple Computer, Inc.",
-    b"agfa": "Agfa Graphics N.V.",
-    b"argl": "ArgyllCMS",  # Not registered
-    b"DCAL": "DisplayCAL",  # Not registered
-    b"bICC": "basICColor GmbH",
-    b"DL&C": "Digital Light & Color",
-    b"EPSO": "Seiko Epson Corporation",
-    b"HDM ": "Heidelberger Druckmaschinen AG",
-    b"HP  ": "Hewlett-Packard",
-    b"KODA": "Kodak",
-    b"lcms": "Little CMS",
-    b"MONS": "Monaco Systems Inc.",
-    b"MSFT": "Microsoft Corporation",
-    b"qato": "QUATOGRAPHIC Technology GmbH",
-    b"XRIT": "X-Rite",
-}
-
-PLATFORM = {
-    b"APPL": "Apple",
-    b"MSFT": "Microsoft",
-    b"SGI ": "Silicon Graphics",
-    b"SUNW": "Sun Microsystems",
-}
-
-PROFILE_CLASS = {
-    b"scnr": "Input device profile",
-    b"mntr": "Display device profile",
-    b"prtr": "Output device profile",
-    b"link": "DeviceLink profile",
-    b"spac": "Color space Conversion profile",
-    b"abst": "Abstract profile",
-    b"nmcl": "Named color profile",
-}
-
-TAGS = {
-    "A2B0": "Device to PCS: Intent 0",
-    "A2B1": "Device to PCS: Intent 1",
-    "A2B2": "Device to PCS: Intent 2",
-    "B2A0": "PCS to device: Intent 0",
-    "B2A1": "PCS to device: Intent 1",
-    "B2A2": "PCS to device: Intent 2",
-    "CIED": "Characterization measurement values",  # Non-standard
-    "DevD": "Characterization device values",  # Non-standard
-    "arts": "Absolute to media relative transform",  # Non-standard (Argyll)
-    "bkpt": "Media black point",
-    "bTRC": "Blue tone response curve",
-    "bXYZ": "Blue matrix column",
-    "chad": "Chromatic adaptation transform",
-    "ciis": "Colorimetric intent image state",
-    "clro": "Colorant order",
-    "cprt": "Copyright",
-    "desc": "Description",
-    "dmnd": "Device manufacturer name",
-    "dmdd": "Device model name",
-    "gamt": "Out of gamut tag",
-    "gTRC": "Green tone response curve",
-    "gXYZ": "Green matrix column",
-    "kTRC": "Gray tone response curve",
-    "lumi": "Luminance",
-    "meas": "Measurement type",
-    "mmod": "Make and model",
-    "ncl2": "Named colors",
-    "pseq": "Profile sequence description",
-    "rTRC": "Red tone response curve",
-    "rXYZ": "Red matrix column",
-    "targ": "Characterization target",
-    "tech": "Technology",
-    "vcgt": "Video card gamma table",
-    "view": "Viewing conditions",
-    "vued": "Viewing conditions description",
-    "wtpt": "Media white point",
-}
-
-TECH = {
-    "fscn": "Film scanner",
-    "dcam": "Digital camera",
-    "rscn": "Reflective scanner",
-    "ijet": "Ink jet printer",
-    "twax": "Thermal wax printer",
-    "epho": "Electrophotographic printer",
-    "esta": "Electrostatic printer",
-    "dsub": "Dye sublimation printer",
-    "rpho": "Photographic paper printer",
-    "fprn": "Film writer",
-    "vidm": "Video monitor",
-    "vidc": "Video camera",
-    "pjtv": "Projection television",
-    "CRT ": "Cathode ray tube display",
-    "PMD ": "Passive matrix display",
-    "AMD ": "Active matrix display",
-    "KPCD": "Photo CD",
-    "imgs": "Photographic image setter",
-    "grav": "Gravure",
-    "offs": "Offset lithography",
-    "silk": "Silkscreen",
-    "flex": "Flexography",
-    "mpfs": "Motion picture film scanner",
-    "mpfr": "Motion picture film recorder",
-    "dmpc": "Digital motion picture camera",
-    "dcpj": "Digital cinema projector",
-}
-
-CIIS = {
-    "scoe": "Scene colorimetry estimates",
-    "sape": "Scene appearance estimates",
-    "fpce": "Focal plane colorimetry estimates",
-    "rhoc": "Reflection hardcopy original colorimetry",
-    "rpoc": "Reflection print output colorimetry",
-}
-
-
-def legacy_PCSLab_dec_to_uInt16(L: float, a: float, b: float) -> list[int]:  # noqa: N802, N803
-    """Convert ICCv2 (legacy) PCS L*a*b* float values to int.
-
-    Only used by LUT16Type and namedColor2Type in ICCv4.
-
-    Args:
-        L (float): L* value in float format.
-        a (float): a* value in float format.
-        b (float): b* value in float format.
-
-    Returns:
-        list[int]: List of int values representing L*, a*, and b* in 16-bit.
-    """
-    return [
-        v * (652.80, 256, 256)[i] + (0, 32768, 32768)[i]
-        for i, v in enumerate((L, a, b))
-    ]
-
-
-def legacy_PCSLab_uInt16_to_dec(  # noqa: N802
-    L_uInt16: int,  # noqa: N803
-    a_uInt16: int,  # noqa: N803
-    b_uInt16: int,  # noqa: N803
-) -> list[float]:
-    """Convert ICCv2 (legacy) PCS L*a*b* to float values.
-
-    Args:
-        L_uInt16 (int): L* value in 16-bit unsigned integer format.
-        a_uInt16 (int): a* value in 16-bit unsigned integer format.
-        b_uInt16 (int): b* value in 16-bit unsigned integer format.
-
-    Returns:
-        list: List of float values representing L*, a*, and b*.
-    """
-    # ICCv2 (legacy) PCS L*a*b* encoding
-    # Only used by LUT16Type and namedColor2Type in ICCv4
-    return [
-        (v - (0, 32768, 32768)[i]) / (65280.0, 32768.0, 32768.0)[i] * (100, 128, 128)[i]
-        for i, v in enumerate((L_uInt16, a_uInt16, b_uInt16))
-    ]
+from DisplayCAL.icc_profile.codecs import (
+    dateTimeNumber,
+    dateTimeNumber_tohex,
+    hexrepr,
+    legacy_PCSLab_dec_to_uInt16,
+    legacy_PCSLab_uInt16_to_dec,
+    s15Fixed16Number,
+    s15Fixed16Number_tohex,
+    s15f16_is_equal,
+    u16Fixed16Number,
+    u16Fixed16Number_tohex,
+    u8Fixed8Number,
+    u8Fixed8Number_tohex,
+    uInt8Number,
+    uInt8Number_tohex,
+    uInt16Number,
+    uInt16Number_tohex,
+    uInt32Number,
+    uInt32Number_tohex,
+    uInt64Number,
+    uInt64Number_tohex,
+    videoCardGamma,
+)
+from DisplayCAL.icc_profile.constants import (
+    CIIS,
+    CMMS,
+    COLOR_PROFILE_SUBTYPE,
+    COLOR_PROFILE_TYPE,
+    COLORANTS,
+    DEBUG,
+    ENC,
+    ENCODINGS,
+    ERROR_PROFILE_NOT_ASSOCIATED_WITH_DEVICE,
+    ERROR_SUCCESS,
+    FS_ENC,
+    GAMUT_VOLUME_ADOBERGB,
+    GAMUT_VOLUME_SMPTE431_P3,
+    GAMUT_VOLUME_SRGB,
+    GEOMETRY,
+    ILLUMINANTS,
+    MANUFACTURERS,
+    OBSERVERS,
+    PLATFORM,
+    PROFILE_CLASS,
+    TAGS,
+    TECH,
+    WCS_PROFILE_MANAGEMENT_SCOPE,
+)
+from DisplayCAL.icc_profile.structures import (
+    ADict,
+    AODict,
+    CRInterpolation,
+    DictList,
+    DictListItem,
+)
 
 
 def create_RGB_A2B_XYZ(  # noqa: N802
@@ -3257,384 +2943,6 @@ def _mp_hdr_tonemap(
     return HDR_XYZ
 
 
-def hexrepr(bytestring: bytes, mapping: None | dict = None) -> str:
-    """Generate hex representation of a bytes instance.
-
-    Args:
-        bytestring (bytes): The bytes to convert.
-        mapping (dict): A dictionary to map the ASCII representation to
-            a string.
-
-    Returns:
-        str: The hex representation of the bytes.
-    """
-    hex_repr = (b"0x%s" % binascii.hexlify(bytestring).upper()).decode()
-    ascii_repr = re.sub(b"[^\x20-\x7e]", b"", bytestring)
-    if ascii_repr == bytestring:
-        hex_repr += f" '{ascii_repr.decode()}'"
-        if mapping:
-            value = mapping.get(ascii_repr)
-            if value:
-                hex_repr = f"{hex_repr} {value}"
-    return hex_repr
-
-
-def dateTimeNumber(binary_string: bytes) -> datetime.datetime:  # noqa: N802
-    """Convert a 12-byte hex representation to a datetime object.
-
-    Byte
-    Offset Content                                     Encoded as...
-    0..1   number of the year (actual year, e.g. 1994) uInt16Number
-    2..3   number of the month (1-12)                  uInt16Number
-    4..5   number of the day of the month (1-31)       uInt16Number
-    6..7   number of hours (0-23)                      uInt16Number
-    8..9   number of minutes (0-59)                    uInt16Number
-    10..11 number of seconds (0-59)                    uInt16Number
-
-    Args:
-        binary_string (bytes): A 12 character long bytes value representing a
-            datetime value.
-
-    Returns:
-        datetime: The datetime object represented by the hex.
-    """
-    Y, m, d, H, M, S = [  # noqa: N806
-        uInt16Number(chunk)
-        for chunk in (
-            binary_string[:2],
-            binary_string[2:4],
-            binary_string[4:6],
-            binary_string[6:8],
-            binary_string[8:10],
-            binary_string[10:12],
-        )
-    ]
-    return datetime.datetime(*(Y, m, d, H, M, S))
-
-
-def dateTimeNumber_tohex(dt: datetime.datetime) -> bytes:  # noqa: N802
-    """Convert a datetime object to a 12-byte hex representation.
-
-    Args:
-        dt (datetime): The datetime object to convert.
-
-    Returns:
-        bytes: The 12-byte hex representation of the datetime.
-    """
-    data = [uInt16Number_tohex(n) for n in dt.timetuple()[:6]]
-    return b"".join(data)
-
-
-def s15Fixed16Number(binaryString: bytes) -> float:  # noqa: N802, N803
-    """Convert a 4-byte hex representation to a float.
-
-    Args:
-        binaryString (bytes): The 4-byte hex representation.
-
-    Returns:
-        float: The number represented by the hex.
-    """
-    return struct.unpack(">i", binaryString)[0] / 65536.0
-
-
-def s15Fixed16Number_tohex(num: int) -> bytes:  # noqa: N802
-    """Convert a number to a 4-byte hex representation.
-
-    Args:
-        num (int): The number to convert.
-
-    Returns:
-        bytes: The 4-byte hex representation of the number.
-    """
-    return struct.pack(">i", round(num * 65536))
-
-
-def s15f16_is_equal(a: bytes, b: bytes, quantizer: None | Callable = None) -> bool:
-    """Compare two s15Fixed16Number values.
-
-    Args:
-        a (bytes): First value.
-        b (bytes): Second value.
-        quantizer (Optional[callable]): A callable to quantize the values.
-            Defaults to None.
-
-    Returns:
-        bool: True if the values are equal, False otherwise.
-    """
-    if quantizer is None:
-
-        def quantizer(v: int) -> float:
-            """Default quantizer for s15Fixed16Number.
-
-            Args:
-                v (int): The value to quantize.
-
-            Returns:
-                float: The quantized value.
-            """
-            return s15Fixed16Number(s15Fixed16Number_tohex(v))
-
-    return colormath.is_equal(a, b, quantizer)
-
-
-def u16Fixed16Number(binaryString: bytes) -> int:  # noqa: N802, N803
-    """Convert a 2-byte hex representation to a number.
-
-    Args:
-        binaryString (bytes): The 2-byte hex representation.
-
-    Returns:
-        int: The number represented by the hex.
-    """
-    return struct.unpack(">I", binaryString)[0] / 65536.0
-
-
-def u16Fixed16Number_tohex(num: int) -> bytes:  # noqa: N802
-    """Convert a number to a 2-byte hex representation.
-
-    Args:
-        num (int): The number to convert.
-
-    Returns:
-        bytes: The 2-byte hex representation of the number.
-    """
-    return struct.pack(">I", round(num * 65536) & 0xFFFFFFFF)
-
-
-def u8Fixed8Number(binaryString: bytes) -> int:  # noqa: N802, N803
-    """Convert a 1-byte hex representation to a number.
-
-    Args:
-        binaryString (bytes): The 1-byte hex representation.
-
-    Returns:
-        int: The number represented by the hex.
-    """
-    return struct.unpack(">H", binaryString)[0] / 256.0
-
-
-def u8Fixed8Number_tohex(num: int) -> bytes:  # noqa: N802
-    """Convert a number to a 1-byte hex representation.
-
-    Args:
-        num (int): The number to convert.
-
-    Returns:
-        bytes: The 1-byte hex representation of the number.
-    """
-    return struct.pack(">H", round(num * 256))
-
-
-def uInt16Number(binaryString: bytes) -> int:  # noqa: N802, N803
-    """Convert a 2-byte hex representation to a number.
-
-    Args:
-        binaryString (bytes): The 2-byte hex representation.
-
-    Returns:
-        int: The number represented by the hex.
-    """
-    return struct.unpack(">H", binaryString)[0]
-
-
-def uInt16Number_tohex(num: int) -> bytes:  # noqa: N802
-    """Convert a number to a 2-byte hex representation.
-
-    Args:
-        num (int): The number to convert.
-
-    Returns:
-        bytes: The 2-byte hex representation of the number.
-    """
-    return struct.pack(">H", round(num))  # num can be float despite the type hint
-
-
-def uInt32Number(binaryString: bytes) -> int:  # noqa: N802, N803
-    """Convert a 4-byte hex representation to a number.
-
-    Args:
-        binaryString (bytes): The 4-byte hex representation.
-
-    Returns:
-        int: The number represented by the hex.
-    """
-    return struct.unpack(">I", binaryString)[0]
-
-
-def uInt32Number_tohex(num: int) -> bytes:  # noqa: N802
-    """Convert a number to a 4-byte hex representation.
-
-    Args:
-        num (int): The number to convert.
-
-    Returns:
-        bytes: The 4-byte hex representation of the number.
-    """
-    return struct.pack(">I", round(num))  # num can be float despite the type hint
-
-
-def uInt64Number(binaryString: bytes) -> int:  # noqa: N802, N803
-    """Convert a 8-byte hex representation to a number.
-
-    Args:
-        binaryString (bytes): The 8-byte hex representation.
-
-    Returns:
-        int: The number represented by the hex.
-    """
-    return struct.unpack(">Q", binaryString)[0]
-
-
-def uInt64Number_tohex(num: int) -> bytes:  # noqa: N802
-    """Convert a number to a 8-byte hex representation.
-
-    Args:
-        num (int): The number to convert.
-
-    Returns:
-        bytes: The 8-byte hex representation of the number.
-    """
-    return struct.pack(">Q", round(num))  # num can be float despite the type hint
-
-
-def uInt8Number(binaryString: bytes) -> int:  # noqa: N802, N803
-    """Convert a 1-byte hex representation to a number.
-
-    Args:
-        binaryString (bytes): The 1-byte hex representation.
-
-    Returns:
-        int: The number represented by the hex.
-    """
-    return struct.unpack(">H", b"\0" + binaryString)[0]
-
-
-def uInt8Number_tohex(num: int) -> bytes:  # noqa: N802
-    """Convert a number to a 1-byte hex representation.
-
-    Args:
-        num (int): The number to convert.
-
-    Returns:
-        bytes: The 1-byte hex representation of the number.
-    """
-    return struct.pack(">H", round(num))[1:2]  # num can be float despite the type hint
-
-
-def videoCardGamma(  # noqa: N802
-    tagData: bytes,  # noqa: N803
-    tagSignature: str,  # noqa: N803
-) -> None | VideoCardGammaTableType | VideoCardGammaFormulaType:
-    """Generate a VideoCardGammaTableType or VideoCardGammaFormulaType tag.
-
-    Args:
-        tagData (bytes): The raw tag data containing the video LUT curves.
-        tagSignature (str): The signature of the tag, typically "vcgt".
-
-    Returns:
-        None | VideoCardGammaTableType | VideoCardGammaFormulaType: The parsed
-            tag data as a VideoCardGammaTableType or VideoCardGammaFormulaType
-            object, or None if the tag type is not recognized.
-    """
-    # reserved = uInt32Number(tagData[4:8])
-    tag_type = uInt32Number(tagData[8:12])
-    if tag_type == 0:  # table
-        return VideoCardGammaTableType(tagData, tagSignature)
-    if tag_type == 1:  # formula
-        return VideoCardGammaFormulaType(tagData, tagSignature)
-    return None
-
-
-class CRInterpolation:
-    """Catmull-Rom interpolation.
-
-    Curve passes through the points exactly, with neighbouring points
-    influencing curvature.points[] should be at least 3 points long.
-
-    Args:
-        points (list[float]): A list of points to interpolate. The list should
-            contain at least 3 points, and each point should be a float value.
-    """
-
-    def __init__(self, points: list[float]) -> None:
-        self.points = points
-
-    def __call__(self, pos: float) -> float:
-        """Interpolate the value at the given position.
-
-        Args:
-            pos (float): The position to interpolate the value for, in the range
-                [0, len(points) - 1].
-
-        Returns:
-            float: The interpolated value at the given position.
-        """
-        lbound = int(math.floor(pos) - 1)
-        ubound = int(math.ceil(pos) + 1)
-        t = pos % 1.0
-        if abs((lbound + 1) - pos) < 0.0001:
-            # sitting on a datapoint, so just return that
-            return self.points[lbound + 1]
-        if lbound < 0:
-            p = self.points[: ubound + 1]
-            # extend to the left linearly
-            while len(p) < 4:
-                p.insert(0, p[0] - (p[1] - p[0]))
-        else:
-            p = self.points[lbound : ubound + 1]
-            # extend to the right linearly
-            while len(p) < 4:
-                p.append(p[-1] - (p[-2] - p[-1]))
-        t2 = t * t
-        return 0.5 * (
-            (2 * p[1])
-            + (-p[0] + p[2]) * t
-            + ((2 * p[0]) - (5 * p[1]) + (4 * p[2]) - p[3]) * t2
-            + (-p[0] + (3 * p[1]) - (3 * p[2]) + p[3]) * (t2 * t)
-        )
-
-
-class ADict(dict):
-    """Convenience class for dictionary key access via attributes.
-
-    Instead of writing aodict[key], you can also write aodict.key
-    """
-
-    def __getattr__(self, name: str) -> Any:  # noqa: ANN401
-        """Get the attribute with the given name.
-
-        Args:
-            name (str): The name of the attribute to get.
-        """
-        if name in self:
-            return self[name]
-        return self.__getattribute__(name)
-
-    def __setattr__(self, name: str, value: Any) -> None:  # noqa: ANN401
-        """Set the attribute with the given name to the given value.
-
-        Args:
-            name (str): The name of the attribute to set.
-            value (Any): The value to set the attribute to.
-        """
-        self[name] = value
-
-
-class AODict(ADict):
-    """Convenience class for dictionary key access via attributes."""
-
-    def __setattr__(self, name: str, value: Any) -> None:  # noqa: ANN401
-        """Set the attribute with the given name to the given value.
-
-        Args:
-            name (str): The name of the attribute to set.
-            value (Any): The value to set the attribute to.
-        """
-        if name == "_keys":
-            object.__setattr__(self, name, value)
-        else:
-            self[name] = value
-
 
 class LazyLoadTagAODict(AODict):
     """Lazy-load (and parse) tag data on access.
@@ -5891,50 +5199,6 @@ class DateTimeType(ICCProfileTag, datetime.datetime):
         return datetime.datetime.__new__(
             cls, dt.year, dt.month, dt.day, dt.hour, dt.minute, dt.second
         )
-
-
-class DictList(list):
-    """ICC dictType Tag list."""
-
-    def __getitem__(self, key: slice | SupportsIndex) -> Any:  # noqa: ANN401
-        """Get item from list.
-
-        Args:
-            key (slice | SupportsIndex): Key of the item.
-
-        Returns:
-            Any: Value of the item.
-        """
-        for item in self:
-            if item[0] == key:
-                return item
-        raise KeyError(key)
-
-    def __setitem__(self, key: slice | SupportsIndex, value: Any) -> None:  # noqa: ANN401
-        """Set item in list.
-
-        Args:
-            key (slice | SupportsIndex): Key of the item.
-            value (Any): Value of the item.
-        """
-        if not isinstance(value, DictListItem):
-            self.append(DictListItem((key, value)))
-
-
-class DictListItem(list):
-    """ICC dictType Tag item."""
-
-    def __iadd__(self, value: Any) -> Self:  # noqa: ANN401
-        """Add value to the last item in the list.
-
-        Args:
-            value (Any): Value to add.
-
-        Returns:
-            DictListItem: The updated list item.
-        """
-        self[-1] += value
-        return self
 
 
 class DictType(ICCProfileTag, AODict):
